requirements_deployment.txt
# Discord Bot - Cloud Deployment Requirements
# Lightweight dependencies for 24/7 deployment without AI/Music

# Core Discord functionality
discord.py>=2.3.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
aiohttp>=3.8.0
requests>=2.28.0

# Image processing (lightweight)
Pillow>=9.0.0

# Text processing
textblob>=0.17.0
emoji>=2.0.0

# Development and testing (optional for production)
pytest>=7.0.0
pytest-asyncio>=0.21.0
